XArray: Fix xas_reload for multi-index entries
xas_reload() was only checking that the head entry was still at the
head index.  If the entry has been split, that's not enough as there
may be a different entry at the specified index now.

Solve this by checking the slot for the requested index instead of the
head index.

Signed-off-by: Matthew Wilcox (Oracle) <willy@infradead.org>
Signed-off-by: Muhammad Guntur <mochtur@gmail.com>
Signed-off-by: kvsnr113 <kvsnrprojkt113@gmail.com>