fixup "gpu: adreno: remove binning based gpu freqtable"
i already figured out the problem of our gpu node mess,
"qcom,speed-bin" in kona starts from 0 to 3 which is 865, 865+, 865-, and 870.
however there is speed-bin 4 out of nowhere in the middle of other speed-bin.
so remove it and make the speed-bin sequence same as pwrlevels.

This partially reverts commit b57277ebf477d3118fddfd0282d39072fee6501f.

Signed-off-by: kvsnr113 <kvsnrprojkt113@gmail.com>